+++ src/main/java/com/jelly/zzirit/domain/cart/dto/request/CartItemAddRequest.java
@@ -1,0 +1,10 @@
+package com.jelly.zzirit.domain.cart.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class CartItemAddRequest {
+	private Long itemId;
+	private int quantity;
+	private boolean isTimeDeal;
+}
